+++ shared/r2-shared-swiftTests/Publication/Services/PublicationServicesBuilderTests.swift
@@ -23,6 +23,7 @@ struct BarServiceA: BarService {}
 class PublicationServicesBuilderTests: XCTestCase {
     
     private let context = PublicationServiceContext(
+        publication: Weak<Publication>(),
         manifest: Manifest(metadata: Metadata(title: "")),
         fetcher: EmptyFetcher()
     )